Newly updated everything working so far
