+++ src/components/Events.jsx
@@ -1,0 +1,13 @@
+import EventCard from './EventCard';
+
+const Events = ({ events }) => {
+  return (
+    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+      {events.map((event, index) => (
+        <EventCard key={index} event={event} />
+      ))}
+    </div>
+  );
+};
+
+export default Events;
